pyproject.toml
[tool.poetry]
name = "messages-api"
version = "0.1.0"
description = ""
authors = ["ShanThatos <shanthkoka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^2.3.3"
waitress = "^2.1.2"
python-dotenv = "^1.0.0"
psutil = "^5.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
